+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0
+
+* Added a few options such as: strList, itemBuilder, highlightTextStyle
+, normalTextStyle, showPreview
+
+
 ## 0.0.1
 
 * Added alphabet scroll. Need to implement more customization options.